update: move subcomponents to proper module
