+++ src/views/components/ControlsClose/ControlsClose.tsx
