+++ src/views/components/ControlsClose/index.ts
@@ -1,0 +1,1 @@
+export { ControlsClose } from './ControlsClose.tsx'
