+++ src/views/components/GamePageList/IGamePageList.ts
@@ -1,6 +1,6 @@
 import { ComponentProps } from 'react'
 import { Game } from '@data/types'
-import { ISearchControls } from '@views/pages/SearchControls/SearchControls.tsx'
+import { ISearchControls } from '@views/components/SearchControls/SearchControls.tsx'
 
 
 export interface IGamePageList extends ComponentProps<'article'>,
