+++ src/views/components/SearchControls/index.ts
@@ -1,0 +1,1 @@
+export {SearchControls} from './SearchControls.tsx'
