--- src/views/pages/ControlsClose/index.ts
@@ -1,1 +1,0 @@
-export { ControlsClose } from './ControlsClose'
