--- src/views/pages/SearchControls/index.ts
@@ -1,1 +1,0 @@
-export {SearchControls} from './SearchControls'
