+++ src/views/pages/index.ts
@@ -2,4 +2,3 @@ export { ErrorPage } from './ErrorPage'
 export { GamePage } from './GamePage'
 export { SearchPage } from './SearchPage'
 export { HomePage } from './HomePage'
-export { SearchControls } from '@views/pages/SearchControls/SearchControls.tsx'
